pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-issue-manager"
version = "0.1.0"
description = "A comprehensive CLI tool for managing GitHub issues with AI-powered content rewriting"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Hans", email = "hans@cyfrin.io"},
]
keywords = ["github", "issues", "cli", "ai", "openrouter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Bug Tracking",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "click>=8.0.0",
    "rich>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Repository = "https://github.com/hans-cyfrin/git-issue-manage"
Issues = "https://github.com/hans-cyfrin/git-issue-manage/issues"

[project.scripts]
git-issue-manager = "git_issue_manager.cli.main:main"
gim = "git_issue_manager.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/git_issue_manager"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
